Attribute identifier was sent as int but according to the AJP protocol description this should be a Byte. See http://tomcat.apache.org/connectors-doc/ajp/ajpv13a.html#Attributes

+++ src/main/java/org/grumblesmurf/ajp/ForwardRequestMessage.java
@@ -128,10 +128,10 @@ class ForwardRequestMessage
             String value = attribute.b;
 
             if (Constants.COMMON_ATTRIBUTES.containsKey(name)) {
-                writeInt(Constants.COMMON_ATTRIBUTES.get(name));
+            	writeByte(Constants.COMMON_ATTRIBUTES.get(name));
                 writeString(value);
             } else {
-                writeInt(Constants.ATTRIBUTE_GENERIC);
+                writeByte(Constants.ATTRIBUTE_GENERIC);
                 writeString(name);
                 writeString(value);
             }
